+++ EatNow/Views/RestaurantView.m
@@ -122,7 +122,7 @@
     self.cuisine.text = restaurant.cuisineStr;
     self.price.text = restaurant.pricesStr;
     self.rating.text = [NSString stringWithFormat:@"%.1f", restaurant.rating.floatValue];
-    self.reviews.text = [NSString stringWithFormat:@"%lu", restaurant.reviews.integerValue];
+    self.reviews.text = [NSString stringWithFormat:@"%lu", (long)restaurant.reviews.integerValue];
     self.distance.text = [NSString stringWithFormat:@"%.1fkm", restaurant.distance];
 }
